extract MountedNode into file
