+++ addon/components/ecr-switch.ts
@@ -3,7 +3,7 @@ import Component from '@ember/component';
 import layout from '../templates/components/ecr-switch';
 import { computed } from '@ember/object';
 import { RouterState } from 'ember-navigator/-private/routeable';
-import { MountedNode } from 'ember-navigator/-private/mounted-router';
+import { MountedNode } from 'ember-navigator/-private/mounted-node';
 
 export default class EcrSwitch extends Component.extend({
   tagName: null,
